+++ crates/symbolicator/src/services/symbolication.rs
@@ -1267,20 +1267,16 @@ fn symbolicate_stacktrace(
     metrics: &mut StacktraceMetrics,
     signal: Option<Signal>,
 ) -> CompleteStacktrace {
-    let mut stacktrace = CompleteStacktrace {
-        thread_id: thread.thread_id,
-        is_requesting: thread.is_requesting,
-        registers: thread.registers.clone(),
-        frames: vec![],
-    };
+    let mut symbolicated_frames = vec![];
+    let mut unsymbolicated_frames_iter = thread.frames.into_iter().enumerate().peekable();
 
-    for (index, mut frame) in thread.frames.into_iter().enumerate() {
+    while let Some((index, mut frame)) = unsymbolicated_frames_iter.next() {
         match symbolicate_frame(caches, &thread.registers, signal, &mut frame, index) {
             Ok(frames) => {
                 if matches!(frame.trust, FrameTrust::Scan) {
                     metrics.scanned_frames += 1;
                 }
-                stacktrace.frames.extend(frames)
+                symbolicated_frames.extend(frames)
             }
             Err(status) => {
                 // Since symbolication failed, the function name was not demangled. In case there is
@@ -1316,6 +1312,25 @@ fn symbolicate_stacktrace(
                     continue;
                 }
 
+                // Glibc inserts an explicit `DW_CFA_undefined: RIP` DWARF rule to say that `_start`
+                // has no return address.
+                // See https://sourceware.org/git/?p=glibc.git;a=blob;f=sysdeps/x86_64/start.S;h=1b3e36826b8a477474cee24d1c931429fbdf6d8f;hb=HEAD#l59
+                // We do not support this due to lack of breakpad support, and will thus use the
+                // previous rule for RIP, which says to look it up the value on the stack,
+                // resulting in an unmapped garbage frame. We work around this by trimming the
+                // trailing garbage frame on the following conditions:
+                // * it is unmapped (UnknownImage)
+                // * this is the last frame to symbolicate (via peek)
+                // * the previous symbolicated frame is `_start`
+                let is_start =
+                    |frame: &SymbolicatedFrame| frame.raw.function.as_deref() == Some("_start");
+                if status == FrameStatus::UnknownImage
+                    && unsymbolicated_frames_iter.peek().is_none()
+                    && symbolicated_frames.last().map_or(false, is_start)
+                {
+                    continue;
+                }
+
                 metrics.unsymbolicated_frames += 1;
                 match frame.trust {
                     FrameTrust::Scan => {
@@ -1330,7 +1345,7 @@ fn symbolicate_stacktrace(
                     metrics.unmapped_frames += 1;
                 }
 
-                stacktrace.frames.push(SymbolicatedFrame {
+                symbolicated_frames.push(SymbolicatedFrame {
                     status,
                     original_index: Some(index),
                     raw: frame,
@@ -1340,8 +1355,7 @@ fn symbolicate_stacktrace(
     }
 
     // we try to find a base frame among the bottom 5
-    if !stacktrace
-        .frames
+    if !symbolicated_frames
         .iter()
         .rev()
         .take(5)
@@ -1351,7 +1365,7 @@ fn symbolicate_stacktrace(
     }
     // macOS has some extremely short but perfectly fine stacks, such as:
     // `__workq_kernreturn` > `_pthread_wqthread` > `start_wqthread`
-    if stacktrace.frames.len() < 3 {
+    if symbolicated_frames.len() < 3 {
         metrics.short_traces += 1;
     }
 
@@ -1359,7 +1373,12 @@ fn symbolicate_stacktrace(
         metrics.bad_traces += 1;
     }
 
-    stacktrace
+    CompleteStacktrace {
+        thread_id: thread.thread_id,
+        is_requesting: thread.is_requesting,
+        registers: thread.registers,
+        frames: symbolicated_frames,
+    }
 }
 
 #[derive(Debug, Copy, Clone)]